エンドポイント URL に app.url_path_for を使用する様にテストをリファクタリング
