+++ tests/bookmark/test_delete_bookmark.py
@@ -3,6 +3,7 @@ from fastapi.testclient import TestClient
 from src.dao.models.bookmark import BookmarkDao
 from src.dao.models.bookmark_tag import BookmarkTagDao
 from src.dao.models.tag import TagDao
+from src.main import app
 from tests.conftest import TEST_URL, SessionForTest
 
 from ..base import BaseTest
@@ -12,6 +13,9 @@ class TestDeleteBookmark(BaseTest):
     """
     ブックマーク削除のテストクラス
     """
+
+    def api_path(self, hashed_id: str) -> str:
+        return app.url_path_for("delete_bookmark", hashed_id=hashed_id)
 
     def test_delete_normal(
         self,
@@ -28,7 +32,7 @@ class TestDeleteBookmark(BaseTest):
         bookmark_id = bookmarks[0].id
 
         # リクエストの送信
-        response = client.delete(f"/bookmarks/{bookmarks[0].hashed_id}")
+        response = client.delete(self.api_path(bookmarks[0].hashed_id))
 
         # レスポンスの検証
         assert response.status_code == 200
@@ -54,7 +58,7 @@ class TestDeleteBookmark(BaseTest):
         hashed_id = "0" * 64
 
         # リクエストの送信
-        response = client.delete(f"/bookmarks/{hashed_id}")
+        response = client.delete(self.api_path(hashed_id))
 
         # レスポンスの検証
         assert response.status_code == 404
@@ -71,10 +75,10 @@ class TestDeleteBookmark(BaseTest):
         """
         # ハッシュIDの長さ不正
         hashed_id = "0" * 63
-        response = client.delete(f"/bookmarks/{hashed_id}")
+        response = client.delete(self.api_path(hashed_id))
         assert response.status_code == 422
 
         # ハッシュIDのフォーマット不正
         hashed_id = "x" * 64
-        response = client.delete(f"/bookmarks/{hashed_id}")
+        response = client.delete(self.api_path(hashed_id))
         assert response.status_code == 422
